requirements.txt
dotenv
torch
torchvision
requests
python-dotenv
Pillow
tqdm
streamlit